refactor(regional): moved methos to region specific utils for  KSA
